test: PlaceControllerTest 통합
